requirements.txt
streamlit
python-dotenv
openai
vaderSentiment
textblob
langdetect